move gen logic to convo updates
